requirements.txt
wheel>=0.40.0
numpy>=1.24.0,<2.0.0
pandas>=2.2.0
python-dotenv>=1.0.0
google-generativeai>=0.3.0
watchdog>=3.0.0
pyarrow>=12.0.0,<14.0.0
streamlit>=1.32.0
grpcio>=1.59.0
sqlalchemy>=2.0.0
pymysql>=1.1.0
cryptography>=42.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
aiohttp>=3.9.0
tenacity>=8.2.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
black>=24.0.0
isort>=5.13.0
mypy>=1.8.0
